use c99 int types in multiboot.h

+++ arch/x86/boot/multiboot.h
@@ -31,82 +31,84 @@
 
 #ifndef __ASM__
 
+#include <stdint.h>
+
 typedef struct module_info {
-  unsigned long mod_start;
-  unsigned long mod_end;
-  unsigned long string;
-  unsigned long reserved;
+  uint32_t mod_start;
+  uint32_t mod_end;
+  uint32_t string;
+  uint32_t reserved;
 } module_info_t;
 
 typedef struct aout_symbol_table {
-  unsigned long tabsize;
-  unsigned long strsize;
-  unsigned long addr;
-  unsigned long reserved;
+  uint32_t tabsize;
+  uint32_t strsize;
+  uint32_t addr;
+  uint32_t reserved;
 } aout_symbol_table_t;
 
 typedef struct elf_section_header_table {
-  unsigned long num;
-  unsigned long size;
-  unsigned long addr;
-  unsigned long shndx;
+  uint32_t num;
+  uint32_t size;
+  uint32_t addr;
+  uint32_t shndx;
 } elf_section_header_table_t;
 
 typedef struct mmap_info {
-  unsigned long size;
-  unsigned long base_addr_low;
-  unsigned long base_addr_high;
-  unsigned long length_low;
-  unsigned long length_high;
-  unsigned long type;
+  uint32_t size;
+  uint32_t base_addr_low;
+  uint32_t base_addr_high;
+  uint32_t length_low;
+  uint32_t length_high;
+  uint32_t type;
 } mmap_info_t;
 
 typedef struct drive_info {
-  unsigned long size;
-  unsigned long drive_number;
-  unsigned long drive_mode;
-  unsigned long drive_cylinders;
-  unsigned long drive_heads;
-  unsigned long drive_sectors;
-  unsigned short drive_ports;
+  uint32_t size;
+  uint32_t drive_number;
+  uint32_t drive_mode;
+  uint32_t drive_cylinders;
+  uint32_t drive_heads;
+  uint32_t drive_sectors;
+  uint16_t drive_ports;
 } drive_info_t;
 
 typedef struct apm_table {
-  unsigned short version;
-  unsigned short cseg;
-  unsigned long offset;
-  unsigned short cseg_16;
-  unsigned short dseg;
-  unsigned short flags;
-  unsigned short cseg_len;
-  unsigned short cseg_16_len;
-  unsigned short dseg_len;
+  uint16_t version;
+  uint16_t cseg;
+  uint32_t offset;
+  uint16_t cseg_16;
+  uint16_t dseg;
+  uint16_t flags;
+  uint16_t cseg_len;
+  uint16_t cseg_16_len;
+  uint16_t dseg_len;
 } apm_table_t;
 
 typedef struct multiboot_info {
-  unsigned long flags;
-  unsigned long mem_lower;
-  unsigned long mem_upper;
-  unsigned long boot_device;
-  unsigned long cmdline;
-  unsigned long mods_count;
-  unsigned long mods_addr;
+  uint32_t flags;
+  uint32_t mem_lower;
+  uint32_t mem_upper;
+  uint32_t boot_device;
+  uint32_t cmdline;
+  uint32_t mods_count;
+  uint32_t mods_addr;
   union {
     aout_symbol_table_t aout_symbol_table;
     elf_section_header_table_t elf_section_header_table;
   } syms;
-  unsigned long mmap_length;
-  unsigned long mmap_addr;
-  unsigned long drives_length;
-  unsigned long config_table;
-  unsigned long boot_loader_name;
-  unsigned long apm_table;
-  unsigned long vbe_control_info;
-  unsigned long vbe_mode_info;
-  unsigned long vbe_mode;
-  unsigned long vbe_interface_seg;
-  unsigned long vbe_interface_off;
-  unsigned long vbe_interface_len;
+  uint32_t mmap_length;
+  uint32_t mmap_addr;
+  uint32_t drives_length;
+  uint32_t config_table;
+  uint32_t boot_loader_name;
+  uint32_t apm_table;
+  uint32_t vbe_control_info;
+  uint32_t vbe_mode_info;
+  uint32_t vbe_mode;
+  uint32_t vbe_interface_seg;
+  uint32_t vbe_interface_off;
+  uint32_t vbe_interface_len;
 } multiboot_info_t;
 
 #endif /* __ASM__ */
